requirements.txt
google-adk>=1.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pytz>=2024.1
uvicorn>=0.30.0
nest-asyncio>=1.6.0